refactor into new subpackage
