+++ uncertainties/new/__init__.py
@@ -1,0 +1,9 @@
+from uncertainties.new.umath import (
+    add_float_funcs_to_ufloat,
+    add_math_funcs_to_umath,
+    add_ufuncs_to_ufloat,
+)
+
+add_float_funcs_to_ufloat()
+add_math_funcs_to_umath()
+add_ufuncs_to_ufloat()
